Refactor Bundles into a separate file.
